requirements.txt
PyMuPDF>=1.23.0
reportlab>=4.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pywin32>=306
numpy>=1.24.0
opencv-python>=4.8.0
